move non-cli code to lib.rs
